Refractoring of User, Student, Parent, Teacher
Reqests now reqire JSON

+++ databaseService/src/main/java/f54148/adminication/repository/ParentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import f54148.adminication.entity.Parent;
 
-public interface ParentRepository  extends CrudRepository<Parent, Integer>{
+public interface ParentRepository  extends CrudRepository<Parent, Long>{
 
 }
