+++ databaseService/src/main/java/f54148/adminication/repository/TeacherRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import f54148.adminication.entity.Teacher;
 
 
-public interface TeacherRepository extends CrudRepository<Teacher, Integer>{
+public interface TeacherRepository extends CrudRepository<Teacher, Long>{
 
 }
